dbeaver/pro#7844 refactor deps, added socket endpoint
