+++ server/bundles/io.cloudbeaver.server/META-INF/MANIFEST.MF
@@ -11,6 +11,7 @@ Require-Bundle: org.eclipse.core.runtime;visibility:=reexport,
  org.apache.commons.jexl,
  org.jkiss.utils;visibility:=reexport,
  org.jkiss.dbeaver.model.sql;visibility:=reexport,
+ org.jkiss.dbeaver.model.lsp,
  org.jkiss.dbeaver.data.gis,
  org.jkiss.bundle.jetty.server;visibility:=reexport,
  com.google.gson;visibility:=reexport,
