--- server/bundles/io.cloudbeaver.service.lsp/build.properties
@@ -1,6 +1,0 @@
-source.. = src/
-output.. = target/classes/
-bin.includes = .,\
-    META-INF/,\
-    plugin.xml,\
-    schema/
