--- server/bundles/io.cloudbeaver.service.lsp/plugin.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?eclipse version="3.2"?>
-
-<plugin>
-
-</plugin>
