--- server/bundles/io.cloudbeaver.service.lsp/src/io/cloudbeaver/lsp/LSPWebSessionProvider.java
@@ -1,4 +1,0 @@
-package io.cloudbeaver.lsp;
-
-public class LSPWebSessionProvider {
-}
