+++ server/bundles/pom.xml
@@ -26,7 +26,6 @@
         <module>io.cloudbeaver.service.rm</module>
         <module>io.cloudbeaver.service.rm.nio</module>
         <module>io.cloudbeaver.service.data.transfer</module>
-        <module>io.cloudbeaver.service.lsp</module>
 
         <module>io.cloudbeaver.service.security</module>
         <module>io.cloudbeaver.service.ldap.auth</module>
