+++ server/features/io.cloudbeaver.server.feature/feature.xml
@@ -39,7 +39,6 @@
     <plugin id="io.cloudbeaver.service.fs" version="0.0.0"/>
     <plugin id="io.cloudbeaver.service.rm" version="0.0.0"/>
     <plugin id="io.cloudbeaver.service.rm.nio" version="0.0.0"/>
-    <plugin id="io.cloudbeaver.service.lsp" version="0.0.0"/>
 
     <plugin id="io.cloudbeaver.service.security" version="0.0.0"/>
     <plugin id="io.cloudbeaver.service.ldap.auth" version="0.0.0"/>
